+++ packages/website/src/pages/data-validation.tsx
@@ -1,7 +1,7 @@
 import TabItem from "@theme/TabItem";
 import Tabs from "@theme/Tabs";
 import { FluentLayout } from "../components/fluent-layout";
-import { Section } from "../components/homepage/section/section";
+import { Section } from "../components/section/section";
 import { SectionedLayout } from "../components/sectioned-layout/sectioned-layout";
 import { UseCaseOverview } from "../components/use-case-overview/use-case-overview";
 import { Links } from "../constants";
