+++ packages/website/src/pages/tooling.tsx
@@ -1,6 +1,6 @@
 import { CodeBlock } from "../components/code-block/code-block";
 import { FluentLayout } from "../components/fluent-layout";
-import { Section } from "../components/homepage/section/section";
+import { Section } from "../components/section/section";
 import { SectionedLayout } from "../components/sectioned-layout/sectioned-layout";
 import { UseCaseOverview } from "../components/use-case-overview/use-case-overview";
 import style from "./tooling.module.css";
